Remove locks when merge() operations are complete
This removes all git notes representing locks for a particular
repository once merge() operations are complete.

In order to accomplish this, this utilises a lockTable present
in the requests object to traverse through all locks that were
acquired during the lifetime of the request operation.

Whilst we're here, augment the removeNotes command to include an
additional parameter for the git object to point to. Also cleanup lock
usage further by making these functions private, to effectively abstract
lock mechanisms away from the user and putting lock creation behind a
factory pattern.